requirements.txt
langchain
langchain-mistralai
langchain-community
langchain-core
langgraph
python-dotenv
requests
duckduckgo-search
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
fastapi
uvicorn[standard]
python-multipart
